pyproject.toml
[tool.poetry]
name = "homework-29-pd12"
version = "0.1.0"
description = ""
authors = ["StasHunter92 <stashunter92@gmail.com>"]
readme = "README.md"
packages = [{include = "homework_29_pd12"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
